w3c-reports/12092014/efl/runned-tests/XMLHttpRequest/send-authentication-basic-cors-not-enabled-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/XMLHttpRequest/resources/auth1/auth.py. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.

FAIL XMLHttpRequest: send() - "Basic" authenticated CORS requests with user name and password passed to open() (asserts failure) assert_throws: function "function () { client.send(null); }" threw object "Error: NETWORK_ERR: XMLHttpRequest Exception 101" that is not a DOMException NetworkError: property "code" is equal to 101, expected 19(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:682:23
http://localhost:8800/XMLHttpRequest/send-authentication-basic-cors-not-enabled.htm:21:22
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/XMLHttpRequest/send-authentication-basic-cors-not-enabled.htm:14:11)
